refactor: ErrorProvider에서 useError를 파일로 분리
