+++ client/src/hooks/useError/useError.tsx
@@ -1,0 +1,12 @@
+import {useContext} from 'react';
+
+import {ErrorContext, ErrorContextType} from './ErrorProvider';
+
+// 에러 컨텍스트를 사용하는 커스텀 훅
+export const useError = (): ErrorContextType => {
+  const context = useContext(ErrorContext);
+  if (!context) {
+    throw new Error('useError must be used within an ErrorProvider');
+  }
+  return context;
+};
